Refactor `PlanStorageBar` to TS so typescript is happy
